refactor(monitor): always pass CustomNotificationTemplate
If we don't pass `CustomNotificationTemplate`, the backend generates
a default notification subjects for alerts in V2 api, but it is not
passed back on POST response even if subsequents GETs will have it,
complicating the state management.

If we pass `CustomNotificationTemplate` with empty strings, POSTS and
GETs will be consistent and the same default notification subjects will
be generated anyway at alert triggering time.

This is also the same technique used by the UI.

Also, do not require anymore to add a subject when defining a
`CustomNotificationTemplate`.