Add secure-onboarding organization acc test
Change summary:
----------------
Terraform acceptance tests use real Terraform configurations to
exercise the code in real plan, apply, refresh, and destroy life cycles.

TF acceptance tests for secure org onboarding need an actual existing gcp project
along with an actual service_principal_key to scrape all folders and projects under the org.
Without it POST /organizations API fail with 500 error and hence we could not
add org acceptance tests working with the sysdig backend APIs.

To handle this, adding the org acceptance tests with skip on this particular error check.
With this, we can get some decent test coverage for the HCL and the APIs for org.