feat!: deprecate cloudsec resources (#671)
This pull request deprecates the resource responsible for making
`cloudsec` calls in the Sysdig Terraform provider.

Specifically, it removes the `sysdig_secure_cloud_account` resource and
its related implementation, tests, and documentation. The changes
include deleting the resource file, its test file, and the corresponding
documentation, as well as cleaning up references in the provider’s
registration and internal client code. This is part of a broader effort
to phase out legacy `cloudsec` resource management from the provider.